problem with path of qr in the product of forms

+++ climmob/products/forms/celerytasks.py
@@ -40,7 +40,13 @@ def createDocumentForm(
         nameOutput += "_" + code
 
     pathqr = os.path.join(path, "qr")
-    os.makedirs(pathqr)
+    if form == "Registration":
+        pathqr = os.path.join(path, "qr")
+
+        if os.path.exists(pathqr):
+            sh.rmtree(pathqr)
+
+        os.makedirs(pathqr)
 
     pathoutput = os.path.join(path, "outputs")
     if not os.path.exists(pathoutput):
@@ -49,10 +55,11 @@ def createDocumentForm(
     PATH2 = os.path.dirname(os.path.abspath(__file__))
     doc = DocxTemplate(PATH2 + "/template/word_template.docx")
     imgsOfQRs = []
-    for package in packages:
+    if form == "Registration":
+        for package in packages:
 
-        qr = create_qr(package, projectid, pathqr)
-        imgsOfQRs.append(InlineImage(doc, qr, width=Mm(50)))
+            qr = create_qr(package, projectid, pathqr)
+            imgsOfQRs.append(InlineImage(doc, qr, width=Mm(50)))
 
     data = {
         "tittle": _(form + " form for the project"),
@@ -68,8 +75,14 @@ def createDocumentForm(
     }
 
     doc.render(data)
+
+    if os.path.exists(pathoutput + "/" + nameOutput + "_" + projectid + ".docx"):
+        print("Lo elimina")
+        os.remove(pathoutput + "/" + nameOutput + "_" + projectid + ".docx")
+
     doc.save(pathoutput + "/" + nameOutput + "_" + projectid + ".docx")
 
-    sh.rmtree(pathqr)
+    if os.path.exists(pathqr):
+        sh.rmtree(pathqr)
 
     return ""
